requirements.txt
botbuilder-integration-aiohttp>=4.14.0
azure-messaging-webpubsubservice
websockets
python-dotenv